+++ frontend/locales/en/common.ts
@@ -105,6 +105,7 @@ const common = {
     "not active": "not active",
 
     // Communication strings
+    "communication": "communication",
     "create communication template": "create communication template",
     "template placeholder": "Write your template here",
     subject: "subject",
@@ -113,6 +114,7 @@ const common = {
     "open in mail application": "open in mail application",
     "choose template": "choose your template",
     "create template": "or create your own",
+    "edit template": "edit the template",
     "additional communication info": "additional info",
     "communication content": "content of the communication",
     medium: "medium",
